requirements.txt
fastapi
uvicorn
python-multipart
requests
twilio
sarvamai
google-generativeai
python-dotenv
